+++ NEWS.md
@@ -1,8 +1,9 @@
 
-# survParamSim 0.1.6
+# survParamSim 0.1.6 (in development)
 
 ## Minor changes
 
+* Fix error when the same variable is included in both `group` and `trt` for `calc_km_pi()`
 
 # survParamSim 0.1.5